Faults: attachments, fixes #10

+++ src/main/java/cz/svjis/servlet/cmd/FaultReportingAttachmentDeleteCmd.java
@@ -5,6 +5,8 @@
  */
 package cz.svjis.servlet.cmd;
 
+import cz.svjis.bean.FaultReport;
+import cz.svjis.bean.FaultReportAttachment;
 import cz.svjis.bean.FaultReportDAO;
 import cz.svjis.servlet.CmdContext;
 import cz.svjis.servlet.Command;
@@ -25,9 +27,8 @@ public class FaultReportingAttachmentDeleteCmd extends Command {
     public void execute() throws Exception {
 
         String parId = getRequest().getParameter("id");
-        String parFaultReportId = getRequest().getParameter("faultReportId");
         
-        if (!validateInput(parId, parFaultReportId)) {
+        if (!validateInput(parId)) {
             RequestDispatcher rd = getRequest().getRequestDispatcher("/InputValidationError.jsp");
             rd.forward(getRequest(), getResponse());
             return;
@@ -36,22 +37,21 @@ public class FaultReportingAttachmentDeleteCmd extends Command {
         FaultReportDAO faultDao = new FaultReportDAO(getCnn());
 
         int id = Integer.parseInt(parId);
-        int faultId = Integer.parseInt(parFaultReportId);
-        faultDao.deleteFaultAttachment(id);
-        String url = "Dispatcher?page=faultDetail&id=" + faultId;
+        FaultReportAttachment fa = faultDao.getFaultReportAttachment(id);
+        FaultReport f = faultDao.getFault(getCompany().getId(), fa.getFaultReportId());
+        if ((f != null) && (!f.isClosed()) && (fa.getUser().getId() == getUser().getId())) {
+            faultDao.deleteFaultAttachment(id);
+        }
+        String url = "Dispatcher?page=faultDetail&id=" + fa.getFaultReportId();
         getRequest().setAttribute("url", url);
         RequestDispatcher rd = getRequest().getRequestDispatcher("/_refresh.jsp");
         rd.forward(getRequest(), getResponse());
     }
     
-    private boolean validateInput(String parId, String parFaultReportId) {
+    private boolean validateInput(String parId) {
         boolean result = true;
         
         if (!Validator.validateInteger(parId, 0, Validator.maxIntAllowed)) {
-            result = false;
-        }
-        
-        if (!Validator.validateInteger(parFaultReportId, 0, Validator.maxIntAllowed)) {
             result = false;
         }
